requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain
langchain-openai
langchain-community
faiss-cpu
sentence-transformers
numpy
pdfplumber
python-docx
pdf2image
PyPDF2
pytesseract
chardet
tqdm
razdel
python-multipart
pandas
openpyxl
python-pptx
docxtpl
huggingface_hub[hf_xet]